+++ typescript/infra/scripts/agent-utils.ts
@@ -1,3 +1,4 @@
+import debug from 'debug';
 import path from 'path';
 import yargs from 'yargs';
 
@@ -10,12 +11,14 @@ import {
   CoreConfig,
   MultiProvider,
   RpcConsensusType,
+  chainMetadata,
   collectValidators,
 } from '@hyperlane-xyz/sdk';
 import { ProtocolType, objMap, promiseObjAll } from '@hyperlane-xyz/utils';
 
 import { Contexts } from '../config/contexts';
-import { environments } from '../config/environments';
+import { agents } from '../config/environments/agents';
+import { validatorBaseConfigsFn } from '../config/environments/utils';
 import { getCurrentKubernetesContext } from '../src/agents';
 import { getCloudAgentKey } from '../src/agents/key-utils';
 import { CloudAgentKey } from '../src/agents/keys';
@@ -28,6 +31,8 @@ import { fetchProvider } from '../src/config/chain';
 import { EnvironmentNames, deployEnvToSdkEnv } from '../src/config/environment';
 import { Role } from '../src/roles';
 import { assertContext, assertRole, readJSON } from '../src/utils/utils';
+
+const debugLog = debug('infra:scripts:utils');
 
 export enum Modules {
   // TODO: change
@@ -120,6 +125,14 @@ export function withKeyRoleAndChain<T>(args: yargs.Argv<T>) {
     .alias('i', 'index');
 }
 
+// missing chains are chains needed which are not as part of defaultMultisigConfigs in sdk/src/consts/ but are in chainMetadata
+export function withMissingChains<T>(args: yargs.Argv<T>) {
+  return args
+    .describe('newChains', 'new chains to add')
+    .string('newChains')
+    .alias('n', 'newChains');
+}
+
 export function assertEnvironment(env: string): DeployEnvironment {
   if (EnvironmentNames.includes(env)) {
     return env as DeployEnvironment;
@@ -129,39 +142,98 @@ export function assertEnvironment(env: string): DeployEnvironment {
   );
 }
 
-export function getEnvironmentConfig(environment: DeployEnvironment) {
-  return environments[environment];
-}
-
-export async function getConfigsBasedOnArgs(argv?: {
+// not requiring to build coreConfig to get agentConfig
+export async function getAgentConfigsBasedOnArgs(argv?: {
   environment: DeployEnvironment;
   context: Contexts;
+  newChains: string;
 }) {
-  const { environment, context = Contexts.Hyperlane } = argv
-    ? argv
-    : await withContext(getArgs()).argv;
-  const envConfig = getEnvironmentConfig(environment);
-  const agentConfig = getAgentConfig(context, envConfig);
-  return { envConfig, agentConfig, context, environment };
+  const {
+    environment,
+    context = Contexts.Hyperlane,
+    newChains,
+  } = argv ? argv : await withMissingChains(withContext(getArgs())).argv;
+
+  const newValidatorCounts: ChainMap<number> = {};
+  if (newChains) {
+    const chains = newChains.split(',');
+    for (const chain of chains) {
+      const [chainName, newValidatorCount] = chain.split('=');
+      newValidatorCounts[chainName] = parseInt(newValidatorCount, 10);
+    }
+  }
+
+  const agentConfig = getAgentConfig(context, environment);
+  // check if new chains are needed
+  const missingChains = checkIfValidatorsArePersisted(agentConfig);
+
+  // if you include a chain in chainMetadata but not in the aw-multisig.json, you need to specify the new chain in new-chains
+  for (const chain of missingChains) {
+    if (!Object.keys(newValidatorCounts).includes(chain)) {
+      throw new Error(`Missing chain ${chain} not specified in new-chains`);
+    }
+    const baseConfig = {
+      [Contexts.Hyperlane]: [],
+      [Contexts.ReleaseCandidate]: [],
+      [Contexts.Neutron]: [],
+    };
+    // supplementing with dummy addresses for validator as part of missingChains
+    const validatorsConfig = validatorBaseConfigsFn(environment, context);
+
+    const validators = validatorsConfig(
+      {
+        ...baseConfig,
+        [context]: Array(newValidatorCounts[chain]).fill('0x0'),
+      },
+      chain as Chains,
+    );
+    // the hardcoded fields are not strictly necessary to be accurate for create-keys.ts
+    // ideally would still get them from the chainMetadata
+    if (!agentConfig.validators) {
+      throw new Error('AgentConfig does not have validators');
+    }
+
+    agentConfig.validators.chains[chain] = {
+      interval: chainMetadata[chain].blocks?.estimateBlockTime ?? 1, // dummy value
+      reorgPeriod: chainMetadata[chain].blocks?.reorgPeriod ?? 0, // dummy value
+      validators,
+    };
+  }
+
+  return {
+    agentConfig,
+    context,
+    environment,
+  };
 }
 
 // Gets the agent config of a specific context.
+// without fetching environment config
 export function getAgentConfig(
   context: Contexts,
-  environment: EnvironmentConfig | DeployEnvironment,
+  environment: DeployEnvironment,
 ): RootAgentConfig {
-  const coreConfig =
-    typeof environment == 'string'
-      ? getEnvironmentConfig(environment)
-      : environment;
-  const agentConfig = coreConfig.agents[context];
-  if (!agentConfig)
-    throw Error(
-      `Invalid context ${context} for environment, must be one of ${Object.keys(
-        coreConfig.agents,
-      )}.`,
+  const agentsForEnvironment = agents[environment] as Record<
+    Contexts,
+    RootAgentConfig
+  >;
+  if (!Object.keys(agents[environment]).includes(context)) {
+    throw new Error(
+      `Context ${context} does not exist in agents for environment ${environment}`,
     );
-  return agentConfig;
+  }
+  return agentsForEnvironment[context];
+}
+
+// check if validators are persisted in agentConfig
+export function checkIfValidatorsArePersisted(
+  agentConfig: RootAgentConfig,
+): Set<ChainName> {
+  const supportedChainNames = agentConfig.contextChainNames.validator;
+  const persistedChainNames = Object.keys(agentConfig.validators?.chains || {});
+  return new Set(
+    supportedChainNames.filter((x) => !persistedChainNames.includes(x)),
+  );
 }
 
 export function getKeyForRole(
@@ -171,8 +243,8 @@ export function getKeyForRole(
   role: Role,
   index?: number,
 ): CloudAgentKey {
-  const environmentConfig = environments[environment];
-  const agentConfig = getAgentConfig(context, environmentConfig);
+  debugLog(`Getting key for ${role} role`);
+  const agentConfig = getAgentConfig(context, environment);
   return getCloudAgentKey(agentConfig, role, chain, index);
 }
 
@@ -185,17 +257,25 @@ export async function getMultiProviderForRole(
   // TODO: rename to consensusType?
   connectionType?: RpcConsensusType,
 ): Promise<MultiProvider> {
+  debugLog(`Getting multiprovider for ${role} role`);
   if (process.env.CI === 'true') {
+    debugLog('Returning multiprovider with default RPCs in CI');
     return new MultiProvider(); // use default RPCs
   }
   const multiProvider = new MultiProvider(txConfigs);
   await promiseObjAll(
     objMap(txConfigs, async (chain, _) => {
-      const provider = await fetchProvider(environment, chain, connectionType);
-      const key = getKeyForRole(environment, context, chain, role, index);
-      const signer = await key.getSigner(provider);
-      multiProvider.setProvider(chain, provider);
-      multiProvider.setSigner(chain, signer);
+      if (multiProvider.getProtocol(chain) === ProtocolType.Ethereum) {
+        const provider = await fetchProvider(
+          environment,
+          chain,
+          connectionType,
+        );
+        const key = getKeyForRole(environment, context, chain, role, index);
+        const signer = await key.getSigner(provider);
+        multiProvider.setProvider(chain, provider);
+        multiProvider.setSigner(chain, signer);
+      }
     }),
   );
 
@@ -212,6 +292,7 @@ export async function getKeysForRole(
   index?: number,
 ): Promise<ChainMap<CloudAgentKey>> {
   if (process.env.CI === 'true') {
+    debugLog('No keys to return in CI');
     return {};
   }