+++ typescript/sdk/src/consts/chains.ts
@@ -17,6 +17,8 @@ export enum Chains {
   fuji = 'fuji',
   gnosis = 'gnosis',
   goerli = 'goerli',
+  inevm = 'inevm',
+  injective = 'injective',
   lineagoerli = 'lineagoerli',
   mantapacific = 'mantapacific',
   moonbasealpha = 'moonbasealpha',
@@ -71,6 +73,8 @@ export const Mainnets: Array<CoreChainName> = [
   Chains.base,
   Chains.scroll,
   Chains.polygonzkevm,
+  Chains.injective,
+  Chains.inevm,
   Chains.viction,
   // Chains.solana,
 ];