+++ typescript/infra/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hyperlane-xyz/infra
 
+## 3.6.2
+
+### Patch Changes
+
+- @hyperlane-xyz/helloworld@3.6.2
+- @hyperlane-xyz/sdk@3.6.2
+- @hyperlane-xyz/utils@3.6.2
+
 ## 3.6.1
 
 ### Patch Changes